+++ packages/pigeon/lib/objc_generator.dart
@@ -92,9 +92,21 @@ const Map<String, String> _propertyTypeForDartTypeMap = <String, String>{
   'Map': 'strong',
 };
 
-String? _objcTypePtrForPrimitiveDartType(String type) {
-  return _objcTypeForDartTypeMap.containsKey(type)
-      ? '${_objcTypeForDartTypeMap[type]} *'
+/// Converts a [List] of [TypeArgument]s to a comma separated [String] to be
+/// used in objc code.
+String _flattenTypeArguments(String? classPrefix, List<TypeArgument> args) {
+  return args
+      .map((TypeArgument e) => e.typeArguments == null
+          ? '${_objcTypeForDartType(classPrefix, e.dataType)} *'
+          : '${_objcTypeForDartType(classPrefix, e.dataType)}<${_flattenTypeArguments(classPrefix, e.typeArguments!)}> *')
+      .reduce((String value, String element) => '$value, $element');
+}
+
+String? _objcTypePtrForPrimitiveDartType(String? classPrefix, Field field) {
+  return _objcTypeForDartTypeMap.containsKey(field.dataType)
+      ? field.typeArguments == null
+          ? '${_objcTypeForDartTypeMap[field.dataType]} *'
+          : '${_objcTypeForDartTypeMap[field.dataType]}<${_flattenTypeArguments(classPrefix, field.typeArguments!)}> *'
       : null;
 }
 
@@ -121,8 +133,8 @@ void _writeClassDeclarations(
   for (final Class klass in classes) {
     indent.writeln('@interface ${_className(prefix, klass.name)} : NSObject');
     for (final Field field in klass.fields) {
-      final HostDatatype hostDatatype = getHostDatatype(
-          field, classes, enums, _objcTypePtrForPrimitiveDartType,
+      final HostDatatype hostDatatype = getHostDatatype(field, classes, enums,
+          (Field x) => _objcTypePtrForPrimitiveDartType(prefix, x),
           customResolver: enumNames.contains(field.dataType)
               ? (String x) => _className(prefix, x)
               : (String x) => '${_className(prefix, x)} *');
@@ -228,38 +240,39 @@ void _writeHostApiDeclaration(Indent indent, Api api, ObjcOptions options) {
   indent.writeln('@protocol $apiName');
   for (final Method func in api.methods) {
     final String returnTypeName =
-        _objcTypeForDartType(options.prefix, func.returnType);
+        _objcTypeForDartType(options.prefix, func.returnType.dataType);
     if (func.isAsynchronous) {
-      if (func.returnType == 'void') {
-        if (func.argType == 'void') {
+      if (func.returnType.dataType == 'void') {
+        if (func.argType.dataType == 'void') {
           indent.writeln(
               '-(void)${func.name}:(void(^)(FlutterError *_Nullable))completion;');
         } else {
           final String argType =
-              _objcTypeForDartType(options.prefix, func.argType);
+              _objcTypeForDartType(options.prefix, func.argType.dataType);
           indent.writeln(
               '-(void)${func.name}:(nullable $argType *)input completion:(void(^)(FlutterError *_Nullable))completion;');
         }
       } else {
-        if (func.argType == 'void') {
+        if (func.argType.dataType == 'void') {
           indent.writeln(
               '-(void)${func.name}:(void(^)($returnTypeName *_Nullable, FlutterError *_Nullable))completion;');
         } else {
           final String argType =
-              _objcTypeForDartType(options.prefix, func.argType);
+              _objcTypeForDartType(options.prefix, func.argType.dataType);
           indent.writeln(
               '-(void)${func.name}:(nullable $argType *)input completion:(void(^)($returnTypeName *_Nullable, FlutterError *_Nullable))completion;');
         }
       }
     } else {
-      final String returnType =
-          func.returnType == 'void' ? 'void' : 'nullable $returnTypeName *';
-      if (func.argType == 'void') {
+      final String returnType = func.returnType.dataType == 'void'
+          ? 'void'
+          : 'nullable $returnTypeName *';
+      if (func.argType.dataType == 'void') {
         indent.writeln(
             '-($returnType)${func.name}:(FlutterError *_Nullable *_Nonnull)error;');
       } else {
         final String argType =
-            _objcTypeForDartType(options.prefix, func.argType);
+            _objcTypeForDartType(options.prefix, func.argType.dataType);
         indent.writeln(
             '-($returnType)${func.name}:($argType*)input error:(FlutterError *_Nullable *_Nonnull)error;');
       }
@@ -279,12 +292,14 @@ void _writeFlutterApiDeclaration(Indent indent, Api api, ObjcOptions options) {
       '- (instancetype)initWithBinaryMessenger:(id<FlutterBinaryMessenger>)binaryMessenger;');
   for (final Method func in api.methods) {
     final String returnType =
-        _objcTypeForDartType(options.prefix, func.returnType);
-    final String callbackType = _callbackForType(func.returnType, returnType);
-    if (func.argType == 'void') {
+        _objcTypeForDartType(options.prefix, func.returnType.dataType);
+    final String callbackType =
+        _callbackForType(func.returnType.dataType, returnType);
+    if (func.argType.dataType == 'void') {
       indent.writeln('- (void)${func.name}:($callbackType)completion;');
     } else {
-      final String argType = _objcTypeForDartType(options.prefix, func.argType);
+      final String argType =
+          _objcTypeForDartType(options.prefix, func.argType.dataType);
       indent.writeln(
           '- (void)${func.name}:($argType*)input completion:($callbackType)completion;');
     }
@@ -401,20 +416,20 @@ void _writeHostApiSource(Indent indent, ObjcOptions options, Api api) {
               '[channel setMessageHandler:^(id _Nullable message, FlutterReply callback) ');
           indent.scoped('{', '}];', () {
             final String returnType =
-                _objcTypeForDartType(options.prefix, func.returnType);
+                _objcTypeForDartType(options.prefix, func.returnType.dataType);
             String syncCall;
-            if (func.argType == 'void') {
+            if (func.argType.dataType == 'void') {
               syncCall = '[api ${func.name}:&error]';
             } else {
               final String argType =
-                  _objcTypeForDartType(options.prefix, func.argType);
+                  _objcTypeForDartType(options.prefix, func.argType.dataType);
               indent.writeln('$argType *input = message;');
               syncCall = '[api ${func.name}:input error:&error]';
             }
             if (func.isAsynchronous) {
-              if (func.returnType == 'void') {
+              if (func.returnType.dataType == 'void') {
                 const String callback = 'callback(wrapResult(nil, error));';
-                if (func.argType == 'void') {
+                if (func.argType.dataType == 'void') {
                   indent.writeScoped(
                       '[api ${func.name}:^(FlutterError *_Nullable error) {',
                       '}];', () {
@@ -429,7 +444,7 @@ void _writeHostApiSource(Indent indent, ObjcOptions options, Api api) {
                 }
               } else {
                 const String callback = 'callback(wrapResult(output, error));';
-                if (func.argType == 'void') {
+                if (func.argType.dataType == 'void') {
                   indent.writeScoped(
                       '[api ${func.name}:^($returnType *_Nullable output, FlutterError *_Nullable error) {',
                       '}];', () {
@@ -445,7 +460,7 @@ void _writeHostApiSource(Indent indent, ObjcOptions options, Api api) {
               }
             } else {
               indent.writeln('FlutterError *error;');
-              if (func.returnType == 'void') {
+              if (func.returnType.dataType == 'void') {
                 indent.writeln('$syncCall;');
                 indent.writeln('callback(wrapResult(nil, error));');
               } else {
@@ -486,15 +501,17 @@ void _writeFlutterApiSource(Indent indent, ObjcOptions options, Api api) {
   indent.addln('');
   for (final Method func in api.methods) {
     final String returnType =
-        _objcTypeForDartType(options.prefix, func.returnType);
-    final String callbackType = _callbackForType(func.returnType, returnType);
+        _objcTypeForDartType(options.prefix, func.returnType.dataType);
+    final String callbackType =
+        _callbackForType(func.returnType.dataType, returnType);
 
     String sendArgument;
-    if (func.argType == 'void') {
+    if (func.argType.dataType == 'void') {
       indent.write('- (void)${func.name}:($callbackType)completion ');
       sendArgument = 'nil';
     } else {
-      final String argType = _objcTypeForDartType(options.prefix, func.argType);
+      final String argType =
+          _objcTypeForDartType(options.prefix, func.argType.dataType);
       indent.write(
           '- (void)${func.name}:($argType*)input completion:($callbackType)completion ');
       sendArgument = 'input';
@@ -512,7 +529,7 @@ void _writeFlutterApiSource(Indent indent, ObjcOptions options, Api api) {
       indent.dec();
       indent.write('[channel sendMessage:$sendArgument reply:^(id reply) ');
       indent.scoped('{', '}];', () {
-        if (func.returnType == 'void') {
+        if (func.returnType.dataType == 'void') {
           indent.writeln('completion(nil);');
         } else {
           indent.writeln('$returnType * output = reply;');